feat: let user choose breath volume
Add command for breath volume, regulation scales patterns
to desired volume.
Add command for cylinder change, regulation handles different
cylinder cross sections.

+++ Core/Inc/Breath.hpp
@@ -7,6 +7,6 @@ struct BreathingPattern {
 	std::array<float, constants::MaxBreathPoints> data;
 	std::size_t length {0};
 	float period {0.0f};
-	float volume {0.0f};
+	int volume {0};
 	float frequency {0.0f};
 };
